+++ packages/stream_feeds/lib/src/state/follow_list_state.dart
@@ -3,6 +3,7 @@ import 'package:state_notifier/state_notifier.dart';
 import 'package:stream_core/stream_core.dart';
 
 import '../models/follow_data.dart';
+import '../models/model_updates.dart';
 import '../models/pagination_data.dart';
 import '../models/query_configuration.dart';
 import 'query/follows_query.dart';
@@ -70,6 +71,17 @@ class FollowListStateNotifier extends StateNotifier<FollowListState> {
     final updatedFollows = state.follows.where((it) {
       return it.id != followId;
     }).toList();
+
+    state = state.copyWith(follows: updatedFollows);
+  }
+
+  /// Handles updates to multiple follows.
+  void onFollowsUpdated(ModelUpdates<FollowData> updates) {
+    final updatedFollows = updates.applyTo(
+      state.follows,
+      key: (it) => it.id,
+      compare: followsSort.compare,
+    );
 
     state = state.copyWith(follows: updatedFollows);
   }